+++ chatserver/mix.exs
@@ -14,8 +14,8 @@ defmodule ChatServer.MixProject do
   # Run "mix help compile.app" to learn about applications.
   def application do
     [
-      mod: {ChatServer, []},
-      extra_applications: [:logger]
+      extra_applications: [:logger],
+      mod: {ChatServer, []}
     ]
   end
